BE - typescript improvements and cleanups

+++ functions/src/service/likes.ts
@@ -1,93 +1,67 @@
 import * as admin from 'firebase-admin'
+import { LikeCreateData } from '../models/Like'
+import db from '../utils/db'
 
-const db = admin.firestore()
+const firebaseDb = admin.firestore()
 
 export const getLikesByUser = async (user_id: string, entity_name?: string) => {
-  let result = db.collectionGroup('likes').where('user_id', '==', user_id)
+  let result = firebaseDb.collectionGroup('likes').where('user_id', '==', user_id)
   if (entity_name) result = result.where('parent_collection_name', '==', entity_name)
   return result.get().then((res) => res.docs.map((doc): any => ({ id: doc.id, ...doc.data() })))
 }
 
 export const getProductLike = async (product_id: string, user_id: string) => {
   const like = await db
-    .collection('products')
-    .doc(product_id)
-    .collection('likes')
+    .getLikes(`products/${product_id}/likes`)
     .doc(`${product_id}_${user_id}_like`)
     .get()
   return like.data()
 }
 
 export const getProductLikes = async (product_id: string) => {
-  const like = await db
-    .collection('products')
-    .doc(product_id)
-    .collection('likes')
-    .get()
-  return like.docs.map((doc): any => ({ ...doc.data(), id: doc.id }))
+  const like = await db.getLikes(`products/${product_id}/likes`).get()
+  return like.docs.map((doc) => ({ ...doc.data(), id: doc.id }))
 }
 
-export const addProductLike = async (product_id: string, user_id: string, data: any = {}) => {
-  const likeRef = db
-    .collection('products')
-    .doc(product_id)
-    .collection('likes')
-    .doc(`${product_id}_${user_id}_like`)
+export const addProductLike = async (product_id: string, user_id: string, data: LikeCreateData) => {
+  const likeRef = db.getLikes(`products/${product_id}/likes`).doc(`${product_id}_${user_id}_like`)
   return await likeRef.set({
     ...data,
     user_id,
     product_id,
-    created_at: new Date(),
+    created_at: FirebaseFirestore.Timestamp.now(),
   })
 }
 
 export const removeProductLike = async (product_id: string, user_id: string) => {
   return await db
-    .collection('products')
-    .doc(product_id)
-    .collection('likes')
+    .getLikes(`products/${product_id}/likes`)
     .doc(`${product_id}_${user_id}_like`)
     .delete()
 }
 
 export const getShopLike = async (shop_id: string, user_id: string) => {
-  const like = await db
-    .collection('shops')
-    .doc(shop_id)
-    .collection('likes')
-    .doc(`${shop_id}_${user_id}_like`)
-    .get()
+  const like = await db.getLikes(`shops/${shop_id}/likes`).doc(`${shop_id}_${user_id}_like`).get()
   return like.data()
 }
 
-export const addShopLike = async (shop_id: string, user_id: string, data: any = {}) => {
-  const likeRef = db
-    .collection('shops')
-    .doc(shop_id)
-    .collection('likes')
-    .doc(`${shop_id}_${user_id}_like`)
+export const addShopLike = async (shop_id: string, user_id: string, data: LikeCreateData) => {
+  const likeRef = db.getLikes(`shops/${shop_id}/likes`).doc(`${shop_id}_${user_id}_like`)
   return await likeRef.set({
     ...data,
     user_id,
     shop_id,
-    created_at: new Date(),
+    created_at: FirebaseFirestore.Timestamp.now(),
   })
 }
 
 export const removeShopLike = async (shop_id: string, user_id: string) => {
-  return await db
-    .collection('shops')
-    .doc(shop_id)
-    .collection('likes')
-    .doc(`${shop_id}_${user_id}_like`)
-    .delete()
+  return await db.getLikes(`shops/${shop_id}/likes`).doc(`${shop_id}_${user_id}_like`).delete()
 }
 
 export const getActivityLike = async (activity_id: string, user_id: string) => {
   const like = await db
-    .collection('activities')
-    .doc(activity_id)
-    .collection('likes')
+    .getLikes(`activities/${activity_id}/likes`)
     .doc(`${activity_id}_${user_id}_like`)
     .get()
   return like.data()
@@ -95,24 +69,20 @@ export const getActivityLike = async (activity_id: string, user_id: string) => {
 
 export const addActivityLike = async (activity_id: string, user_id: string) => {
   const likeRef = db
-    .collection('activities')
-    .doc(activity_id)
-    .collection('likes')
+    .getLikes(`activities/${activity_id}/likes`)
     .doc(`${activity_id}_${user_id}_like`)
   return await likeRef.set({
     parent_collection_path: 'activities',
     parent_collection_name: 'activities',
     user_id,
     activity_id,
-    created_at: new Date(),
+    created_at: FirebaseFirestore.Timestamp.now(),
   })
 }
 
 export const removeActivityLike = async (activity_id: string, user_id: string) => {
   return await db
-    .collection('activities')
-    .doc(activity_id)
-    .collection('likes')
+    .getLikes(`activities/${activity_id}/likes`)
     .doc(`${activity_id}_${user_id}_like`)
     .delete()
 }
@@ -120,11 +90,7 @@ export const removeActivityLike = async (activity_id: string, user_id: string) =
 // no other way around this except adding the comment_id as field inside "comments" collectionGroup
 export const addCommentLike = async (activity_id: string, comment_id: string, user_id: string) => {
   const likeRef = await db
-    .collection('activities')
-    .doc(activity_id)
-    .collection('comments')
-    .doc(comment_id)
-    .collection('comment_likes')
+    .getLikes(`activities/${activity_id}/comments/${comment_id}/comment_likes`)
     .doc(`${comment_id}_${user_id}_like`)
 
   return await likeRef.set({
@@ -132,7 +98,7 @@ export const addCommentLike = async (activity_id: string, comment_id: string, us
     parent_collection_name: 'comments',
     user_id,
     comment_id,
-    created_at: new Date(),
+    created_at: FirebaseFirestore.Timestamp.now(),
   })
 }
 
@@ -143,11 +109,7 @@ export const removeCommentLike = async (
   user_id: string
 ) => {
   return await db
-    .collection('activities')
-    .doc(activity_id)
-    .collection('comments')
-    .doc(comment_id)
-    .collection('comment_likes')
+    .getLikes(`activities/${activity_id}/comments/${comment_id}/comment_likes`)
     .doc(`${comment_id}_${user_id}_like`)
     .delete()
 }
